Refactor & cleanup service controller
