+++ pkg/controller/service/service_controller_test.go
@@ -129,7 +129,7 @@ func TestCreateExternalLoadBalancer(t *testing.T) {
 
 	for _, item := range table {
 		controller, cloud, client := newController()
-		err, _ := controller.createLoadBalancerIfNeeded("foo/bar", item.service)
+		_, err := controller.createLoadBalancerIfNeeded("foo/bar", item.service)
 		if !item.expectErr && err != nil {
 			t.Errorf("unexpected error: %v", err)
 		} else if item.expectErr && err == nil {
@@ -260,50 +260,6 @@ func TestUpdateNodesInExternalLoadBalancer(t *testing.T) {
 	}
 }
 
-func TestGetNodeConditionPredicate(t *testing.T) {
-	tests := []struct {
-		node         v1.Node
-		expectAccept bool
-		name         string
-	}{
-		{
-			node:         v1.Node{},
-			expectAccept: false,
-			name:         "empty",
-		},
-		{
-			node: v1.Node{
-				Status: v1.NodeStatus{
-					Conditions: []v1.NodeCondition{
-						{Type: v1.NodeReady, Status: v1.ConditionTrue},
-					},
-				},
-			},
-			expectAccept: true,
-			name:         "basic",
-		},
-		{
-			node: v1.Node{
-				Spec: v1.NodeSpec{Unschedulable: true},
-				Status: v1.NodeStatus{
-					Conditions: []v1.NodeCondition{
-						{Type: v1.NodeReady, Status: v1.ConditionTrue},
-					},
-				},
-			},
-			expectAccept: false,
-			name:         "unschedulable",
-		},
-	}
-	pred := getNodeConditionPredicate()
-	for _, test := range tests {
-		accept := pred(&test.node)
-		if accept != test.expectAccept {
-			t.Errorf("Test failed for %s, expected %v, saw %v", test.name, test.expectAccept, accept)
-		}
-	}
-}
-
 // TODO(a-robinson): Add tests for update/sync/delete.
 
 func TestProcessServiceUpdate(t *testing.T) {
@@ -398,7 +354,7 @@ func TestProcessServiceUpdate(t *testing.T) {
 	for _, tc := range testCases {
 		newSvc := tc.updateFn(tc.svc)
 		svcCache := controller.cache.getOrCreate(tc.key)
-		obtErr, retryDuration := controller.processServiceUpdate(svcCache, newSvc, tc.key)
+		retryDuration, obtErr := controller.processServiceUpdate(svcCache, newSvc, tc.key)
 		if err := tc.expectedFn(newSvc, obtErr, retryDuration); err != nil {
 			t.Errorf("%v processServiceUpdate() %v", tc.testName, err)
 		}
@@ -444,7 +400,7 @@ func TestSyncService(t *testing.T) {
 				srv := controller.cache.getOrCreate("external-balancer")
 				srv.state = defaultExternalService()
 			},
-			expectedErr: fmt.Errorf("Service somethingelse not in cache even though the watcher thought it was. Ignoring the deletion."),
+			expectedErr: fmt.Errorf("Service somethingelse not in cache even though the watcher thought it was. Ignoring the deletion"),
 		},
 		*/
 
@@ -506,7 +462,7 @@ func TestProcessServiceDeletion(t *testing.T) {
 			},
 			expectedFn: func(svcErr error, retryDuration time.Duration) error {
 
-				expectedError := "Service external-balancer not in cache even though the watcher thought it was. Ignoring the deletion."
+				expectedError := "Service external-balancer not in cache even though the watcher thought it was. Ignoring the deletion"
 				if svcErr == nil || svcErr.Error() != expectedError {
 					//cannot be nil or Wrong error message
 					return fmt.Errorf("Expected=%v Obtained=%v", expectedError, svcErr)
@@ -580,7 +536,7 @@ func TestProcessServiceDeletion(t *testing.T) {
 		//Create a new controller.
 		controller, cloud, _ = newController()
 		tc.updateFn(controller)
-		obtainedErr, retryDuration := controller.processServiceDeletion(svcKey)
+		retryDuration, obtainedErr := controller.processServiceDeletion(svcKey)
 		if err := tc.expectedFn(obtainedErr, retryDuration); err != nil {
 			t.Errorf("%v processServiceDeletion() %v", tc.testName, err)
 		}
@@ -695,129 +651,5 @@ func TestDoesExternalLoadBalancerNeedsUpdate(t *testing.T) {
 		if obtainedResult != tc.expectedNeedsUpdate {
 			t.Errorf("%v needsUpdate() should have returned %v but returned %v", tc.testName, tc.expectedNeedsUpdate, obtainedResult)
 		}
-	}
-}
-
-//All the testcases for ServiceCache uses a single cache, these below test cases should be run in order,
-//as tc1 (addCache would add elements to the cache)
-//and tc2 (delCache would remove element from the cache without it adding automatically)
-//Please keep this in mind while adding new test cases.
-func TestServiceCache(t *testing.T) {
-
-	//ServiceCache a common service cache for all the test cases
-	sc := &serviceCache{serviceMap: make(map[string]*cachedService)}
-
-	testCases := []struct {
-		testName     string
-		setCacheFn   func()
-		checkCacheFn func() error
-	}{
-		{
-			testName: "Add",
-			setCacheFn: func() {
-				cS := sc.getOrCreate("addTest")
-				cS.state = defaultExternalService()
-			},
-			checkCacheFn: func() error {
-				//There must be exactly one element
-				if len(sc.serviceMap) != 1 {
-					return fmt.Errorf("Expected=1 Obtained=%d", len(sc.serviceMap))
-				}
-				return nil
-			},
-		},
-		{
-			testName: "Del",
-			setCacheFn: func() {
-				sc.delete("addTest")
-
-			},
-			checkCacheFn: func() error {
-				//Now it should have no element
-				if len(sc.serviceMap) != 0 {
-					return fmt.Errorf("Expected=0 Obtained=%d", len(sc.serviceMap))
-				}
-				return nil
-			},
-		},
-		{
-			testName: "Set and Get",
-			setCacheFn: func() {
-				sc.set("addTest", &cachedService{state: defaultExternalService()})
-			},
-			checkCacheFn: func() error {
-				//Now it should have one element
-				Cs, bool := sc.get("addTest")
-				if !bool {
-					return fmt.Errorf("is Available Expected=true Obtained=%v", bool)
-				}
-				if Cs == nil {
-					return fmt.Errorf("CachedService expected:non-nil Obtained=nil")
-				}
-				return nil
-			},
-		},
-		{
-			testName: "ListKeys",
-			setCacheFn: func() {
-				//Add one more entry here
-				sc.set("addTest1", &cachedService{state: defaultExternalService()})
-			},
-			checkCacheFn: func() error {
-				//It should have two elements
-				keys := sc.ListKeys()
-				if len(keys) != 2 {
-					return fmt.Errorf("Elementes Expected=2 Obtained=%v", len(keys))
-				}
-				return nil
-			},
-		},
-		{
-			testName:   "GetbyKeys",
-			setCacheFn: nil, //Nothing to set
-			checkCacheFn: func() error {
-				//It should have two elements
-				svc, isKey, err := sc.GetByKey("addTest")
-				if svc == nil || isKey == false || err != nil {
-					return fmt.Errorf("Expected(non-nil, true, nil) Obtained(%v,%v,%v)", svc, isKey, err)
-				}
-				return nil
-			},
-		},
-		{
-			testName:   "allServices",
-			setCacheFn: nil, //Nothing to set
-			checkCacheFn: func() error {
-				//It should return two elements
-				svcArray := sc.allServices()
-				if len(svcArray) != 2 {
-					return fmt.Errorf("Expected(2) Obtained(%v)", len(svcArray))
-				}
-				return nil
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		if tc.setCacheFn != nil {
-			tc.setCacheFn()
-		}
-		if err := tc.checkCacheFn(); err != nil {
-			t.Errorf("%v returned %v", tc.testName, err)
-		}
-	}
-}
-
-//Test a utility functions as its not easy to unit test nodeSyncLoop directly
-func TestNodeSlicesEqualForLB(t *testing.T) {
-	numNodes := 10
-	nArray := make([]*v1.Node, 10)
-
-	for i := 0; i < numNodes; i++ {
-		nArray[i] = &v1.Node{}
-		nArray[i].Name = fmt.Sprintf("node1")
-	}
-	if !nodeSlicesEqualForLB(nArray, nArray) {
-		t.Errorf("nodeSlicesEqualForLB() Expected=true Obtained=false")
 	}
 }
